+++ contracts/multisig-improved/src/action_types/mod.rs
@@ -1,4 +1,6 @@
 pub mod discard;
 pub mod perform;
+pub mod perform_endpoints;
 pub mod propose;
+pub mod propose_endpoints;
 pub mod sign;
